Share unit conversion helper
